pyproject.toml
[project]
name = "dataq-influxdb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "influxdb-client>=1.50.0",
    "pydantic-settings>=2.13.1",
    "pytest>=9.0.2",
    "python-engineio==3.14.2",
    "python-socketio==4.6.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "websocket-client>=1.9.0",
]
